+++ tfmtcnn/prepare_data/lfw.py
@@ -10,7 +10,7 @@ import tensorflow as tf
 from tfmtcnn.prepare_data import tfrecords
 from tfmtcnn.prepare_data.utils import IoU
 from tfmtcnn.prepare_data.bboxdata import BBox
-from tfmtcnn.prepare_data.Landmark_utils import rotate, flip
+from tfmtcnn.prepare_data.landmarks import rotate, flip
 from tfmtcnn.prepare_data import ioutils
 
 """
